+++ game/scrabble.py
@@ -12,10 +12,17 @@ class Scrabble:
         for _ in range(players_count):
             self.players.append(Player())
         self.current_player = None
-        self.turn = 0
+        self.turn = 1
     def playing(self):
         return True
 
     def next_turn(self):
+        if self.current_player is None:
+            self.current_player = self.players[0]
+        elif self.current_player == self.players[-1]:
+            self.current_player = self.players[0]
+        else:
+            player_turn = self.players.index(self.current_player) + 1
+            self.current_player = self.players[player_turn]
         self.turn += 1